MMSI work
All work squashed into one commit. Individual commit messages below.
Unnecessary files removed.

Code creates a panel and tab interface. In the Reachability test you can add a row to the table.
Test new panel.
Fixed the connections panel to have its own URLs and started working on the mockup of the data
Added mock data to reachability table.
Added a mockapi python class for showing a list of mockobjects and adding to the list
Added ability to mock create, run and delete test
Added update template, form and view along with minor changes to the create view
Added the update and details view to the UI
Added a run test view. Still working on displaying a test results page when the button is clicked in the view
Added status of Pass, Fail and Pending. Added changes to create test to match the source/destination type.
Added status lights to the reachability test table. The lights display a stub list of elements for now. Still need to link them to the details page.
Status lights are working properly and are now showing a list of past time stamps.
Finished working on synching the list of past/fail runs with with the status lights.
Added formatting and links to the list of pass/failed runs displayed on the screen when hover over status light.
Deleted Troubleshoot tab and cleaned some of the code. Also, updated the status lights to have icons.
Change the placement of the status lights and text. Added quick test link to a results page.
Added the dialog to run a quick test and save it if necessary.
Modified mock object to hold all the data from creation. Modified the create reachability test form to capture all the data necessary as well as do some validation on it.
Added formatting for the update test view.
Finished setting up the dialog for the run quick test view.
Finish formatting the dialog box for the save button in the quick test details page.
Started implementing the Network Template tab.
Fixed to final bug in save dialog view. Started working on the layout of the Network Template tab.
Added the basic layout of the network template tab and a dialog box for the Apply Network Template button. Submitting the form does nothing currently.
Added a heat template parser and stored the value in a python object for processing.
Added ability to create form based on the sample file given.
Fixed the dialog view to move smoothy between the different phases of the applying a network template.
Fixed the retrouting issue between dialog boxes. Added code handed to us for template rendering. Started working on the rendering of the template.
Created python object from template, converted it to JSON format, passed it to html template and rendered it. Need to position the bocks correctly on the screen and link them up automatically.
Finish the placement and rendering of the network template on the tab. Added a removed network template button but its still not functional.
Added the remove network template dialog. Fixed a display issue with the links. Formatted the mock command prompt. Removed the reachability test box fromt he network template display.
Added validation for selecting a template. Fix some css format issues. Cleaned up the code. Added comments to help with readability of code. Added a README file for installation instructions. Added a copy of PyYAML installation package.
Added style changes to the mock command line.
Added Big Switch Networks logo to the connections panel. Changed remove network template for remove network instance.

Conflicts:
	openstack_dashboard/dashboards/admin/dashboard.py
	openstack_dashboard/dashboards/project/images/views.py